Expose counter for getnativecontracts (#900)
* RpcServer: unify getnativecontracts and getcontractstate, fix #852

Reuse the same model and have updatecounter in the output. All the other
fields were the same already, so old RpcClient can get results without
any problems.

Signed-off-by: Roman Khimov <roman@nspcc.ru>

* RpcClient: use ContractState for GetNativeContractsAsync

Now it's exactly the same server-side, so can be reused. Notice that
UpdateCounter trick allows to parse responses from old servers (that are
lacking "updatecounter" for this call), this simplifies upgrade, but can be
removed in future versions.

Signed-off-by: Roman Khimov <roman@nspcc.ru>

---------

Signed-off-by: Roman Khimov <roman@nspcc.ru>